+++ src/http/socket.js
@@ -2,7 +2,7 @@
 
 /**
  * Pterodactyl - Daemon
- * Copyright (c) 2015 - 2017 Dane Everitt <dane@daneeveritt.com>.
+ * Copyright (c) 2015 - 2018 Dane Everitt <dane@daneeveritt.com>.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -40,7 +40,11 @@ class WebSocket {
             }
 
             this.server.hasPermission('s:console', params.handshake.query.token, (err, hasPermission) => {
-                if (err || !hasPermission) {
+                if (err) {
+                    return next(new Error('There was an error while attempting to validate your permissions.'));
+                }
+
+                if (!hasPermission) {
                     return next(new Error('You do not have permission to access the socket for this server.'));
                 }