+++ src/index.js
@@ -2,7 +2,7 @@
 
 /**
  * Pterodactyl - Daemon
- * Copyright (c) 2015 - 2017 Dane Everitt <dane@daneeveritt.com>.
+ * Copyright (c) 2015 - 2018 Dane Everitt <dane@daneeveritt.com>.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -30,6 +30,7 @@ const compareVersions = require('compare-versions');
 const Fs = require('fs-extra');
 const _ = require('lodash');
 const Keypair = require('keypair');
+const Getos = require('getos');
 
 const Log = rfr('src/helpers/logger.js');
 const Package = rfr('package.json');
@@ -39,7 +40,7 @@ const Config = new ConfigHelper();
 Log.info('+ ------------------------------------ +');
 Log.info(`| Running Pterodactyl Daemon v${Package.version}    |`);
 Log.info('|        https://pterodactyl.io        |');
-Log.info('|  Copyright 2015 - 2017 Dane Everitt  |');
+Log.info('|  Copyright 2015 - 2018 Dane Everitt  |');
 Log.info('+ ------------------------------------ +');
 Log.info('Loading modules, this could take a few seconds.');
 
@@ -152,16 +153,33 @@ Async.auto({
                 });
             },
             (exists, scall) => {
+                Getos((err, os) => {
+                    scall(err, exists, os);
+                });
+            },
+            (exists, os, scall) => {
                 if (exists) {
                     return scall();
                 }
 
-                let UserCommand = 'adduser --system --no-create-home --shell /bin/false --group';
-                if (!_.isUndefined(process.env.IS_ALPINE) && process.env.IS_ALPINE === 'true') {
-                    UserCommand = `addgroup -S ${Config.get('docker.container.username', 'pterodactyl')} && adduser -S -D -H -G ${Config.get('docker.container.username', 'pterodactyl')} -s /bin/false`;
+                let UserCommand = '';
+                const Username = Config.get('docker.container.username', 'pterodactyl');
+
+                switch (_.get(os, 'dist')) {
+                case 'Alpine Linux':
+                    UserCommand = `addgroup -S ${Username} && adduser -S -D -H -G ${Username} -s /bin/false ${Username}`;
+                    break;
+                case 'Ubuntu Linux':
+                case 'Debian':
+                case 'Centos':
+                    UserCommand = `useradd --system --no-create-home --shell /bin/false ${Username}`;
+                    break;
+                default:
+                    return scall(new Error('Unable to create a pterodactyl user and group, unknown operating system.'));
                 }
-                Proc.exec(`${UserCommand} ${Config.get('docker.container.username', 'pterodactyl')}`, {}, err => {
-                    if (err && (!_.includes(err.message, 'already exists') || !_.includes(err.message, `user '${Config.get('docker.container.username', 'pterodactyl')}' in use`))) {
+
+                Proc.exec(UserCommand, {}, err => {
+                    if (err && (!_.includes(err.message, 'already exists') || !_.includes(err.message, `user '${Username}' in use`))) {
                         return scall(err);
                     }
 
@@ -184,23 +202,23 @@ Async.auto({
             },
         ], callback);
     }],
-    check_services: ['setup_sftp_user', (r, callback) => {
-        Service.boot(callback);
-    }],
-    check_network: ['setup_sftp_user', (r, callback) => {
-        Log.info('Checking container networking environment...');
-        Network.init(callback);
-    }],
-    setup_timezone: ['check_network', (r, callback) => {
+    setup_timezone: ['setup_sftp_user', (r, callback) => {
         Log.info('Configuring timezone file location...');
         Timezone.configure(callback);
     }],
-    setup_network: ['check_network', 'setup_timezone', (r, callback) => {
-        Log.info('Checking pterodactyl0 interface and setting configuration values.');
+    check_network: ['setup_timezone', (r, callback) => {
+        Log.info('Checking container networking environment...');
+        Network.init(callback);
+    }],
+    setup_network: ['check_network', (r, callback) => {
+        Log.info('Ensuring correct network interface for containers...');
         Network.interface(callback);
     }],
+    check_services: ['setup_network', (r, callback) => {
+        Service.boot(callback);
+    }],
     init_servers: ['check_services', (r, callback) => {
-        Log.info('Attempting to load servers and initialize daemon...');
+        Log.info('Beginning server initialization process.');
         Initialize.init(callback);
     }],
     configure_perms: ['init_servers', (r, callback) => {